[Refactor] Move pipeline runtime filter and observer primitives
