+++ be/module_boundary_manifest.json
@@ -514,6 +514,7 @@
         "be/src/exec/runtime_filter/**",
         "be/src/exec/pipeline/primitives/operator_exec_stats.h",
         "be/src/exec/pipeline/primitives/operator_runtime_access.h",
+        "be/src/exec/pipeline/runtime_filter_hub.h",
         "be/src/exec/pipeline/runtime_filter_core_types.h",
         "be/src/exec/pipeline/scan/dynamic_morsel_queue_builder.h",
         "be/src/exec/pipeline/scan/fixed_morsel_queue_builder.h",
@@ -526,6 +527,7 @@
         "exec/pipeline/pipeline_fwd.h",
         "exec/pipeline/operator.h",
         "exec/pipeline/primitives/",
+        "exec/pipeline/runtime_filter_hub.h",
         "exec/pipeline/runtime_filter_core_types.h",
         "exec/pipeline/scan/scan_morsel.h",
         "exec/pipeline/scan/morsel_queue.h",
