+++ be/src/compute_env/CMakeLists.txt
@@ -34,7 +34,6 @@ ADD_BE_LIB(ComputeEnv
     data_stream/local_pass_through_buffer.cpp
     data_stream/sender_queue.cpp
     pipeline/driver_limiter.cpp
-    pipeline/observer.cpp
     result/buffer_control_block.cpp
     result/record_batch_queue.cpp
     result/result_buffer_mgr.cpp
