+++ be/src/compute_env/data_stream/data_stream_recvr.h
@@ -42,9 +42,9 @@
 #include "common/runtime_profile.h"
 #include "common/status.h"
 #include "compute_env/data_stream/local_pass_through_buffer.h"
-#include "compute_env/pipeline/observer.h"
 #include "compute_env/sorting/merge_path.h"
 #include "exec/pipeline/pipeline_fwd.h"
+#include "exec/pipeline/primitives/pipeline_observer.h"
 #include "gen_cpp/Types_types.h" // for TUniqueId
 #include "runtime/descriptors.h"
 #include "runtime/query_statistics.h"
