+++ be/src/exec/pipeline/exchange/exchange_parallel_merge_source_operator.cpp
@@ -17,7 +17,7 @@
 #include "common/config_exec_flow_fwd.h"
 #include "compute_env/data_stream/data_stream_mgr.h"
 #include "compute_env/data_stream/data_stream_recvr.h"
-#include "compute_env/pipeline/observer.h"
+#include "exec/pipeline/primitives/pipeline_observer.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/sort_exec_exprs.h"
 #include "runtime/exec_env.h"
