+++ be/src/exec/pipeline/exchange/multi_cast_local_exchange.h
@@ -19,7 +19,7 @@
 
 #include "column/chunk.h"
 #include "common/runtime_profile.h"
-#include "compute_env/pipeline/observer.h"
+#include "exec/pipeline/primitives/pipeline_observer.h"
 #include "exec/pipeline/source_operator.h"
 #include "exec/spill/dir_manager.h"
 #include "fs/fs.h"
