+++ be/src/exec/pipeline/runtime_filter_types.cpp
@@ -15,6 +15,7 @@
 #include "exec/pipeline/runtime_filter_types.h"
 
 #include "common/config_scan_io_fwd.h"
+#include "exprs/predicate.h"
 #include "hashjoin/hash_joiner_fwd.h"
 #include "runtime/runtime_filter_builder.h"
 #include "runtime/runtime_filter_factory.h"
