+++ be/src/exec/pipeline/scan/olap_scan_context.h
@@ -20,10 +20,10 @@
 
 #include "base/phmap/phmap_fwd_decl.h"
 #include "column/column_access_path.h"
-#include "compute_env/pipeline/observer.h"
 #include "exec/olap_scan_prepare.h"
 #include "exec/pipeline/context_with_dependency.h"
 #include "exec/pipeline/operator.h"
+#include "exec/pipeline/primitives/pipeline_observer.h"
 #include "exec/pipeline/scan/balanced_chunk_buffer.h"
 #include "runtime/global_dict/parser.h"
 #include "runtime/runtime_state_fwd.h"
