+++ be/src/exec/pipeline/schedule/timeout_tasks.h
@@ -14,8 +14,8 @@
 
 #pragma once
 
-#include "compute_env/pipeline/observer.h"
 #include "compute_env/pipeline/pipeline_timer.h"
+#include "exec/pipeline/primitives/pipeline_observer.h"
 #include "runtime/descriptors.h"
 
 namespace starrocks::pipeline {
