+++ be/src/exec/pipeline/sort/partition_sort_sink_operator.h
@@ -21,7 +21,7 @@
 #include "common/statusor.h"
 #include "exec/chunks_sorter.h"
 #include "exec/pipeline/operator_factory.h"
-#include "exec/pipeline/runtime_filter_types.h"
+#include "exec/pipeline/runtime_filter_hub.h"
 #include "exec/pipeline/sort/sort_context.h"
 #include "exec/pipeline/spill_process_channel.h"
 #include "exec/sort_exec_exprs.h"
