+++ be/src/exec/pipeline/source_operator.h
@@ -16,10 +16,10 @@
 
 #include <utility>
 
-#include "compute_env/pipeline/observer.h"
 #include "exec/pipeline/adaptive/adaptive_fwd.h"
 #include "exec/pipeline/operator_factory.h"
 #include "exec/pipeline/pipeline_fwd.h"
+#include "exec/pipeline/primitives/pipeline_observer.h"
 #include "exec/pipeline/scan/chunk_source.h"
 #include "exec/pipeline/scan/morsel.h"
 #include "exec/workgroup/work_group_fwd.h"
