+++ be/src/exec/pipeline/wait_operator.cpp
@@ -18,8 +18,8 @@
 
 #include "base/concurrency/stopwatch.hpp"
 #include "common/config_exec_flow_fwd.h"
-#include "compute_env/pipeline/observer.h"
 #include "exec/pipeline/fragment_context.h"
+#include "exec/pipeline/primitives/pipeline_observer.h"
 #include "exec/pipeline/schedule/timeout_tasks.h"
 
 namespace starrocks::pipeline {
