+++ be/src/exec/pipeline/wait_operator.h
@@ -19,9 +19,9 @@
 
 #include "base/debug/debug_action.h"
 #include "column/vectorized_fwd.h"
-#include "compute_env/pipeline/observer.h"
 #include "exec/limited_pipeline_chunk_buffer.h"
 #include "exec/pipeline/operator_factory.h"
+#include "exec/pipeline/primitives/pipeline_observer.h"
 #include "exec/pipeline/schedule/timeout_tasks.h"
 #include "exec/pipeline/source_operator.h"
