+++ be/src/runtime/lookup_stream_mgr.h
@@ -26,9 +26,9 @@
 #include "bthread/mutex.h"
 #include "column/vectorized_fwd.h"
 #include "common/global_types.h"
-#include "compute_env/pipeline/observer.h"
 #include "exec/pipeline/lookup_request.h"
 #include "exec/pipeline/pipeline_fwd.h"
+#include "exec/pipeline/primitives/pipeline_observer.h"
 #include "gen_cpp/Types_types.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "runtime/runtime_fwd.h"
